+++ NEWS.md
@@ -1,7 +1,12 @@
-# mapvizieR 0.0.2 | 07/29/2016
+# NYSEDtools 0.0.3 | 07/29/2016
 
-* added `sirs301_import_csvs`, a csv import function for SIRS 301 tested/not tested report data
+* constructor function for `sirs301` object, and print method.
+* added contributor code of conduct.
 
-# mapvizieR 0.0.1 | 07/29/2016
+# NYSEDtools 0.0.2 | 07/29/2016
+
+* added `sirs301_import_csvs`, a csv import function for SIRS 301 tested/not tested report data.
+
+# NYSEDtools 0.0.1 | 07/29/2016
 
 * hello, world!